module of results is optimized
