+++ js/data/data.js
@@ -3,7 +3,7 @@ import audioData from './audioData';
 const initialState = {
   level: 0,
   lives: 3,
-  time: 300,
+  time: 10,
   answers: []
 };
