+++ js/screens/result/result.js
@@ -1,29 +1,15 @@
 import ResultView from '../../views/result-view';
 import {displayScreen} from '../../lib/screenRender';
 import showWelcome from '../welcome/welcome';
-import {stateGame, arrayResults} from '../../data/data';
-import countPoints from '../../lib/countPoints';
+import {stateGame} from '../../data/data';
 
-const result = (currentResult) => new ResultView(currentResult);
+const result = new ResultView();
+result.replayHandler = () => {
+  displayScreen(showWelcome().element);
+};
 
 result.init = () => {
-  const counterPoints = countPoints(stateGame.answers, stateGame.lives);
-  const currentResult = {
-    points: counterPoints.points,
-    numberOfQuickAnswers: counterPoints.numberOfQuickAnswers,
-    lives: stateGame.lives,
-    time: stateGame.timer.time,
-    timeLeft: stateGame.time - stateGame.timer.time,
-    id: arrayResults.length + 1
-  };
-
-  const showResult = result(currentResult);
-  showResult.replayHandler = () => {
-    displayScreen(showWelcome().element);
-    window.clearInterval(stateGame.timer.id);
-  };
-
-  displayScreen(showResult.element);
+  window.clearInterval(stateGame.timer.id);
 };
 
 export default () => result;
